requirements.txt
flask
beautifulsoup4
requests
dateparser
uuid
pymongo[srv]
gunicorn
python-dotenv
